Create sparsejac package
