+++ pyproject.toml
@@ -1,0 +1,56 @@
+[project]
+
+name = "sparsejac"
+version = "0.0.0"
+description = "Efficient forward- and reverse-mode sparse Jacobians using Jax."
+keywords = ["jax", "jacobian", "sparse"]
+readme = "README.md"
+requires-python = ">=3.7"
+license = {file = "LICENSE"}
+
+authors = [
+    {name = "Martin Schubert", email = "mfschubert@gmail.com"},
+]
+maintainers = [
+    {name = "Martin Schubert", email = "mfschubert@gmail.com"},
+]
+
+dependencies = [
+    "jax",
+    "networkx",
+    "numpy",
+    "scipy",
+]
+
+[project.optional-dependencies]
+tests = [
+    "parameterized",
+    "pytest",
+]
+dev = [
+    "bump-my-version",
+    "darglint",
+    "sparsejac[tests]",
+    "pre-commit",
+]
+
+[build-system]
+requires = ["setuptools>=45", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.black]
+line-length = 88
+target-version = ['py310']
+
+[tool.mypy]
+python_version = "3.10"
+strict = true
+
+[tool.isort]
+multi_line_output = 3
+line_length = 88
+include_trailing_comma = true
+
+[tool.pydocstyle]
+inherit = false
+convention = "google"
